maps and power data
had to add google maps lib to the project.properties.  seems redundent as libs are linked in the header.  Also had to add back in some power vars like V1 and I1 which I had removed because the pin on the original harness broke, so they were no longer available.  But the new harness came wired up, so no pins to break.